Adding simplified gpmsa api
This API feels more light-weight.  There's fewer objects that need
instantiating.  Everything is essentially in one GaussianProcessHelper object.
All the simulation and experimental data is in there, and I also envisage
having the GaussianProcessHelper object be a subclass of a ModelEmulatorHelper
object of some kind so in future releases we can support PCE.  Class names are
up for discussion.  Perhaps 'helper' is a bad name.  Maybe GaussianProcessData
instead?