Update observation handling to use GslVector
In particular, this updates the implementation of lnValue to use some of the
baked methods in GslVector to make computation of the misfit and the norm
cleaner.  This will also be beneficial when it comes to implementing the case
for full observation error covariance matrix, since then we can use the canned
methods that solve linear systems.