Fix bug in MLE compute and add test
All that was happening was a NULL pointer dereference when a particular chain
didn't contain a state that maximised the likelihood (probability 1 event).

Now unifiedPositionsOfMaximum will only fill up a variable on process
zero containing all the states that attain maxima.  This variable on
other processes will be a vector of length zero, hence the removal of
the checks on subSequenceSize == 0.

Also added a test for the unifiedPositionsOfMaximum method in
SequenceOfVectors.