Guard against odd process configuration.
It's possible to be running with multiple processes in the fullComm()
but only 1 subenvironment, like if the user forgets to update the number
of subenvironments in the input file. This is a strange configuration
for the interpolation builder to deal with and would result in
redundant work, so we just won't support it.