Reworked sync_data() algorithm.
There was still one lingering bug for the parallel evaluation of
the interpolation values: Allgather is meant for equal numbers
of values per process. In our case, this may not be true. So, we
need to use the gatherv version. The other change is instead of
doing an allgatherv, we just do a gatherv to processor 0, then
broadcast to all other processes. We were doing a broadcast
anyway, so instead of doing a many-to-many collective, we use only
point-to-many collectives.

The example now works for me for an artitrary number of subenvironments.