Make MetropolisHastingsSG use new parser
This required an update to one of the regression tests because we no longer
print the read options from the base class (because of a lack of prefix).
However, this is useful for debugging so I'd be happy to see this come back in
future hygeine sweeps of the input option reading functionality.