First draft of interpolation surrogate reader
Will read simple ASCII format for values of interpolation surrogate.
Idea is to read in file written by a forthcoming write method.
Currently, we still require the InterpolationSurrogateData object
to have been constructed prior to the reading, so while the data
is there in the input file to contruct that data, we'd need to refactor
the builder a bit, so at this point it's just easier to double
check the data is consistent.