Define ScopedPtr
The main idea is to wrap a smart pointer for forward/backward compatibility.
We needs this because: 1. unique_ptr (natural choice) is C++11 and
QUESO is not requiring C++11, 2. quto_ptr is deprecated, and
3. We'd like to remove the Boost dependency someday. So this uses
a template typedef idiom to wrap them up and provide what we need
based on what's available. Order of preference is: std::unique_ptr,
boost::scoped_ptr, std::auto_ptr. Right now, we don't test for C++11
so std::unique_ptr will never be found at this point. But, we require
Boost so (unless something really bad happened, and we'll get a configure
error anyway since we require boost), we'll always have scoped_ptr.
I put auto_ptr in there for future work. To make std::unique_ptr
a possibility in the future, we need to test/enable C++11 in the compiler
and then add an autoconf macro that tests for unique_ptr and then
AC_SUBST(HAVE_CXX11_UNIQUE_PTR). Similarly, for std::quto_ptr, we
need to define and autoconf macro to test for std::auto_ptr and
AC_SUBST(HAVE_CXX_AUTO_PTR).