Add InterpolationSurrogateDataSet
This is a simple container class to support multiple
InterpolationSurrogateData objects that have the same parameter
domain. This is useful for generalizing the building of interpolation
surrogates to evaluate multiple quantities simultaneously. For example,
if the user is solving a PDE, but will evaluate multiple quantities
based on that PDE solution.