Updated InterpolationSurrogate example data writing
Now checks if the user had more than 1 process running and, if so,
will use the second process to write the second data set. Example
went from 41 seconds to 24 seconds on my workstation; without writing
the data, it runs in less than 1 second, so all the time is spent
writing.