Make tests pass without MPI
This adds strongly typed wrappers around MPI communication routines.  They are
strongly typed so that with or without MPI, we can work out the size of the
type in memory that is being communicated.  In the case where we don't have
MPI, or have MPI but the communicator has size 1, we resort to a std::memcpy.