Fix issue involving unified MLE on non-inter0Comm
This patch adds some safeguards to avoid code that does inter0Comm computing
from being executed by processes that don't belong to the inter0Comm.
Specifically, computing the unified MLE (and positions) and computing
correlation matrices are to only be done on the inter0Comm, since this is where
all the samples live anyway.