Adding HDF5 output support for subchains
This relies on an awful chainSize * numParams *COPY*.  This is completely and
utterly the wrong thing to do, but I wanted to get something working.  Most of
the logic here was actually copied from the unified HDF5 write, which doesn't
seem to be working at the time of this commit.