Display CPPFLAGS with queso-config, configure
Internally we use QUESO_CPPFLAGS because somehow CPPFLAGS doesn't
get propagated through to Automake. This required a s/QUESO_CPPFLAGS/AM_CPPFLAGS
in the examples/Makefile.am. Now, everyone should be respecting
QUESO_CPPFLAGS in the main, build, examples, and tests. We also
now see it in the configure summary.