Adding TransitionKernelFactory subclass
Regarding the factory_map method, we explicitly specialise on the GslVector
and GslMatrix types.  For other vector and matrix types, we'll need to
copy-pasta this method and instantiate for those other vector and matrix types.

Fortunately, the plan is to remove the vector and matrix template types so we
feel that this is the best short-term solution.  Once the vector and matrix
template types are removed, this will become a nonissue.