Implement different factories for different ctors
The "stochastic newton" tk ctor and the "random walk" tk ctor take different
arguments.

This patch implements different factories for these different ctors and then
also templates the factories around the derived types so that the user can
implement their own TK without havign to subclass a factory for a type of ctor
that is already supported.