Update C++11 assert macros
These use the strategy employed by @roystgnr in the Antioch PR
libantioch/antioch#213. This required explicitly stating std::string
on many of the strings used in the queso_require_*_msg macros
for the cases to compile.

Also make sure C++ thinks numbers are a float-type. For the new
style of C++11 asserts, we do checks for casts of
both arguments. In these cases, the second test was failing because
it was casting the double to an int and, therefore, 0 > 0 wasn't
true anymore.