Rejig build system to make ANN build out of source
We update the m4 macro to nuke the ANN_LIBS var (sets it to an empty
string) since the assumption is that the build system figures it out by
relying on libtool.

We update the src/Makefile.am to make queso only conditionally include
the ANN_CFLAGS and the libANN.la file.

The src/contrib/ANN/Makefile.am contains updates to build ANN using
libtool (and not install it).

The ANN test has been disabled for now.