Make Environment use ScopedPtr for m_optionsObj
This required a little dance.  The user can pass in a pointer to their
own instantiation of EnvOptionsValues for setting options
programatically.  Unfortunately, if QUESO just throws this into a
ScopedPtr then the BaseEnvironment dtor will delete it.  If the user's
object is also destructed on their side, we get a double delete.

So, to solve this issue we decided just to copy the users object.  This
is probably more well-behaved anyway.  It guarantees the users's object
doesn't change since if we do decide to do any modifications, they will
be done on the copy and not on the user's object.

"But, but, copies are expensive!" I hear you cry.  That's quite alright
because an EnvOptionsValues object is really not too big.  It's probably
around O(100) strings, ints, and doubles, and the copy is only done once
during the construction of the environment.