Update SIP docs
This just clarifies that the users' options are copied.  This means the
user can call delete on their options, but also that if they change
their object after a StatisticalInverseProblem object has been
constructed then these changes will have no effect on the behaviour of
of the instantiated StatisticalInverseProblem object.