Clean up MetropolisHasingsSG documentation
This just clarifies that the users' options are copied.  This means the
user can call delete on their options, but also that if they change
their object after a MetropolisHasingsSG object has been
constructed then these changes will have no effect on the behaviour of
of the instantiated MetropolisHasingsSG object.