Remove instantiation of global tk factory vars
These have been replaced by local instantiations in a
TKFactoryInitializer that is constructed by MetropolisHastingsSG.

The reason we do this is because if the user wants to link statically
with QUESO, the default behaviour of ld is to say, "I see you are not
using these global tk factory variables, so I'll remove them for you."
This is a feature of ld.  It is stripping out symbols the user doesn't
use.  ld can't possibly know that it shouldn't strip out those symbols
just because the user relies on the side effect of the symbols being
created.

Go figure.

Forcing the linker to see that the symbols are used by QUESO means they
won't be stripped when linking statically.

Also moves all build_tk method implementations to header

The user will need these for their custom TKs.  And QUESO needs them as
all the instantiations of global TK factory registrations have been
removed.