Add VectorRV::has_realizer() test
Otherwise we have no way to test for a realizer without throwing an
exception.

It's probably *more correct* to just catch the exception and use that
as our test, but I really like being able to "catch throw" in gdb to
jump straight to errors, and a redundant method doesn't seem like a
high price to pay to preserve that.