License stamping only needs to be in top Makefile
Before, we (redundantly) had the license stamping in src/Makefile.am
and test/Makefile.am. This keeps it in only the top-level Makefile.am
since we sniff all the files want to stamp at configure time. This
way, it's only in one place and there's only 1 turd file to clean up.