Add ability to set components of finite diff vec
Adds a setStepSize(index, value) method to BaseScalarFunction to allow
the user to programatically set individual elements of the finite
difference step size vector in the case of a multi-dimensional scalar
function.

The ability to modify inidividual elements of the finite difference step
size vector via the input file is not supported.

The user may also pass a scalar (via the input file or programatically)
and if the function is multi-dimensional then the scalar is used to set
every component of the finite difference step size vector to that scalar
value.