Fix up dregs of old lnValue API usage
Mainly in BayesianJointPdf and in ScalarFunctionSynchronizer.

I had to update some of the tests too.  It turns out that users who were
already providing gradients with the old API all of a sudden had their
implementations replaced with a finite difference.

There's a way around this behaviour that's hacky, but it's probably not
worth it unless there are serious complaints.