Don't let the sampler overwrite the dirty iter
We store the iteration at which the user set the dirty flag to true.
Logic has been updated to stop the sampler from overwriting this
iteration.  Previously it was overwriting every iteration an adaptation
wasn't done because adaptation was the thing responsible for setting the
dirty flag back to false.  Now the adaptation just relies on the dirty
iteration tracker.