Fix serial env failure when trilinos enabled
The fix was to wrap around Epetra's abstract base class communicator
(Eptera_Comm) which handles both the case where the user wants mpi
(Epetra_MpiComm) and when the user doesn't have MPI (Epetra_SerialComm).

I had to resort to a dynamic cast for the Comm() method because that
needs to return an honest-to-God MPI communicator type.  Perhaps a
better solution is to #ifdef around that method only when MPI is
available.