Make covMatrixIsDirty/cleanCovMatrix pure virtual
Now the onus is on the client to implement and return the correct thing
when they modify the covariance matrix and subsequently have to clean
it.

It basically amounts to the same thing.  The client will have to carry
around an internal bool flag or something, but making these pure virtual
means that it's impossible for the client to rely on default behaviour
and they're now required to think about how modifying TK properties
affects AM.

Update test to use internal bool flag