Put cov matrix transform logic into TK
This transform logic happens when calling updateCovMatrix on the TK to
feed it a new cov matrix after Adaptive Metropolis.

Since the sequence generator holds a reference to the base class, I had
to add a pure virtual updateCovMatrix method in the base class.  Derived
classes that don't have covariance matrices or don't support adaptivity
will throw an exception in this method.