Add TensorProductMesh interpolation code
I tried to make this as general as possible.  The benefits of this
design are that 1-D problems aren't slowed down by 4-D interpolation
code and that 5-D problems will be trivial to support if we ever need.

The cost of this design is that it's much harder to understand than a
couple quadruply-nested loops would have been.  I should probably add
some cppunit tests...