Fix discrepancy correlation strength indexing
This fix just applies to the indexing in the hyperparameter domain.
Updates to the test and the corresponding regression test solutions were
also made in this commit.

The test indexing needed to be updated because it assumed there was a
single discprepancy correlation strength parameter.  In actuality, there
are two, because numOutputs*dimScenario = 2 * 1 = 2.