Only add emulator data precision to emulator part
In the scalar case this'll boil down to nothing; that is we add the
inverse of the emulator data precision to every diagonal element.  In
the multivariate case we skip over the
num_emulator_bases*num_experiments piece that doesn't correspond to the
emulator part of the covariance matrix.

This commit also updates a regression test solution.  But not the other
one.  Not sure why.